Get rid or more -Wconversion warnings.
git-svn-id: https://svn.r-project.org/R/trunk@55439 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/deparse.c
@@ -453,7 +453,7 @@ SEXP attribute_hidden do_dump(SEXP call, SEXP op, SEXP args, SEXP rho)
             for (i = 0, nout = 0; i < nobjs; i++)
             {
                 const char *s;
-                int extra = 6;
+                unsigned int extra = 6;
                 if (CAR(o) == R_UnboundValue)
                     continue;
                 SET_STRING_ELT(outnames, nout++, STRING_ELT(names, i));
@@ -1336,7 +1336,7 @@ static void writeline(LocalParseData *d)
 
 static void print2buff(const char *strng, LocalParseData *d)
 {
-    int tlen, bufflen;
+    size_t tlen, bufflen;
 
     if (d->startline)
     {
